Put status data from bot in info box
Under the infoBox object in the index.html file, a new attribute ".botStatus" is added and filled with the JSON data returned from the bot. 

This data is then used to bill an info innerText display showing the ping distance, D3, D4, D5 pin levels, and the status, e.g. "driving", "idle", "playback"

This attribute is available to use in javascript programs in robot.js or even in javascript snippets in index.html