Add 2 piece body for sensors
This svg file describes a 2 part body which is simple to assemble based on a design by brotherrobot.com. The two parts fold around each other and attach with zip ties avoiding the need for hot glue. The back end of the top section can be held down by pinning it through the sides with skewers or the clipped off ends of the zip ties, IF the cardboard is cut so that the channels in it run left right instead of front to back.  The holes in the front accommodate a ping sensor (now supported in firmware), and a ball point pen barrel for an LED IR transmitter (not yet supported in firmware). Those will plug into a standard small "whiteboard" / "breadboard" setting on top of the batteries inside. If you shift the breadboard to the side, it can accommodate both.